move many methods into batch.py
